+++ resdk/constants.py
@@ -9,3 +9,5 @@ ReSDK constants.
 """
 
 CHUNK_SIZE = 8000000  # 8MB
+
+RESOLWE_DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S.%f'